feat: add hungarian language
