+++ packages/shared/package.json
@@ -1,5 +1,5 @@
 {
   "name": "@activepieces/shared",
-  "version": "0.10.52",
+  "version": "0.10.53",
   "type": "commonjs"
 }
