+++ packages/shared/src/lib/common/index.ts
@@ -1,2 +1,3 @@
 export * from './utils'
 export * from './base-model'
+export * from './locale'
