+++ packages/shared/src/lib/common/locale.ts
@@ -1,0 +1,14 @@
+export enum LocalesEnum {
+    DUTCH = 'nl',
+    ENGLISH = 'en',
+    GERMAN = 'de',
+    ITALIAN = 'it',
+    FRENCH = 'fr',
+    HUNGARIAN = 'hu',
+    SPANISH = 'es',
+    JAPANESE = 'ja',
+    INDONESIAN = 'id',
+    VIETNAMESE = 'vi',
+    CHINESE_SIMPLIFIED = 'zh',
+    PORTUGUESE = 'pt',
+}
